+++ src/main.js
@@ -11,8 +11,6 @@ Vue.config.productionTip = false
 
 Vue.component("downloadExcel", JsonExcel);
 
-import mixins from './mixins/global'
-Vue.mixin(mixins)
 // let app;
 // firebase.auth().onAuthStateChanged(user => {
 //   console.log(user);
@@ -22,7 +20,8 @@ Vue.mixin(mixins)
 // });
 
 let app;
-firebase.auth().onAuthStateChanged(() => {
+firebase.auth().onAuthStateChanged(user => {
+  console.log('this is the authenticated user: ' + user);
   if(!app) {
     app = new Vue({
         router,
