+++ src/views/Clients.vue
@@ -1,72 +1,57 @@
 <template>
   <div class="clients">
+    <h1 class="subtitle-1 grey--text">Clients</h1>
     <v-container class="my-5">
-      <v-row>
-        <v-col cols="6" md="10">
-            <h2 v-if="$vuetify.breakpoint.mdAndUp" class="info--text">Clients</h2>
-            <h4 v-if="$vuetify.breakpoint.smAndDown" class="info--text">Clients</h4>
-        </v-col>
-        <v-col cols="6" md="2" class="text-right">
-          <CreateClient />
-        </v-col>
-      </v-row>
-
-      <v-row>
-        <v-col cols="12">
-          <ClientsList />
-        </v-col>
-      </v-row>
-
-    <v-container class="my-5">
-      
+      <CreateClient />
       <v-layout row class="mb-3">
-
-
-
+        <v-btn small text @click="sortBy('companyName')">
+          <v-icon left small>mdi-folder</v-icon>
+          <span class="caption text-lowercase">By Client name</span>
+        </v-btn>
+        <v-btn small text @click="sortBy('vat')">
+          <v-icon left small>mdi-folder</v-icon>
+          <span class="caption text-lowercase">By VAT</span>
+        </v-btn>
       </v-layout>
-      <!-- <v-card class="px-3 py-0 rounded-0" v-for="client in clients" :key="client.companyName" @click.stop="details(client.id)">
+      <v-card class="px-3 py-0 rounded-0" v-for="client in clients" :key="client.companyName" @click.stop="details(client.id)">
         <v-layout row wrap :class="`pa-3 client ${client.status}`">
           <v-flex xs12 md6>
             <div class="caption grey--text pa-0">Company Name</div>
-            <div class="subtitle-2">{{ client.client_info.companyName }}</div>
+            <div class="subtitle-2">{{ client.companyName }}</div>
           </v-flex>
           <v-flex xs6 sm4 md2>
             <div class="caption grey--text">VAT</div>
-            <div class="subtitle-2">{{ client.client_info.vat }}</div>
+            <div class="subtitle-2">{{ client.vat }}</div>
           </v-flex>
           <v-flex xs6 sm4 md2>
             <div class="caption grey--text">BULSTAT</div>
-            <div class="subtitle-2">{{ client.client_info.bulstat }}</div>
+            <div class="subtitle-2">{{ client.bulstat }}</div>
           </v-flex>
           <v-flex xs6 sm4 md2>           
             <div class="float-right">
               <div class="caption grey--text">Status</div>
-              <v-chip small :class="`${client.client_info.status} white--text caption`"> {{ client.client_info.status }} </v-chip>
-               
+              <v-chip small :class="`${client.status} white--text caption`"> {{ client.status }} </v-chip>
+               <!-- <button size="sm" @click.stop="details(client.id)">Details</button> -->
             </div>
           </v-flex>
         </v-layout>
-      </v-card> -->
+      </v-card>
 
     </v-container>
-        </v-container>
   </div>
 </template>
 
 <script>
 import CreateClient from '../components/CreateClient'
-import ClientsList from '../components/Clients/ClientsList'
-// import db from '@/fb'
+import db from '@/fb'
 import router from '../router'
 
 export default {
-  
-  components: { CreateClient, ClientsList },
+  components: { CreateClient },
   data: () => ({
+    clients: [],
   }),
-  
   created() {
-    
     // db.collection('users').onSnapshot(res => {
     //   const changes = res.docChanges();
     //   changes.forEach(change => {
@@ -79,17 +64,17 @@ export default {
     //       }
     //   })
     // })
-    //   db.collectionGroup('client-info').onSnapshot(res => {
-    //   const changes = res.docChanges();
-    //   changes.forEach(change => {
-    //       if (change.type === 'added'){
-    //         this.clients.push({
-    //           ...change.doc.data(),
-    //           id: change.doc.id
-    //         });
-    //       }
-    //   })
-    // })
+      db.collectionGroup('client-info').onSnapshot(res => {
+      const changes = res.docChanges();
+      changes.forEach(change => {
+          if (change.type === 'added'){
+            this.clients.push({
+              ...change.doc.data(),
+              id: change.doc.id
+            });
+          }
+      })
+    })
   },
   methods: {
     sortBy(prop){
