+++ src/views/Documents.vue
@@ -2,24 +2,13 @@
   <div class="documents">
     <v-container class="my-5">
       <v-row>
-        <v-col cols="12">
-            <h2 v-if="$vuetify.breakpoint.mdAndUp" class="info--text">
-              <span v-if="type == 'all'">All Documents</span>
-              <span v-if="type != 'all'">Documents / <span v-if="type == 'monthly'">{{getMonth(month)}}</span></span>
-              <span v-if="type == 'yearly'">{{year}}</span></h2>
-            <h4 v-if="$vuetify.breakpoint.smAndDown" class="info--text">
-              <span v-if="type == 'All'">All Documents</span>
-              <span v-if="type != 'all'">Documents / <span v-if="type == 'monthly'">{{getMonth(month)}}</span>
-              </span> <span v-if="type == 'yearly'">{{year}}</span></h4>
-            <div class="d-flex">
-              <p :class="type == 'all' ? 'active info--text pl-2 link' : 'info--text pl-2 link'" @click="type = 'all'">All Documents</p>
-              <p :class="type == 'monthly' ? 'active info--text pl-2 link' : 'info--text pl-2 link'" @click="type = 'monthly'">Monthly Documents</p>
-              <p :class="type == 'yearly' ? 'active info--text pl-2 link' : 'info--text pl-2 link'" @click="type = 'yearly'">Yearly Statements</p>
-            </div>
+        <v-col cols="8">
+            <h2 v-if="$vuetify.breakpoint.mdAndUp" class="info--text">Monthly Documents</h2>
+            <h4 v-if="$vuetify.breakpoint.smAndDown" class="info--text">Monthly Documents</h4>
         </v-col>
-        <!-- <v-col cols="4" class="text-right" v-if="!loading">
+        <v-col cols="4" class="text-right" v-if="!loading">
             <v-btn depressed color="info" @click="addModal = !addModal">Upload Document</v-btn>
-        </v-col> -->
+        </v-col>
       </v-row>
 
       <v-row v-if="loading" class="my-5">
@@ -33,248 +22,204 @@
         </v-col>
       </v-row>
 
-      <div v-if="type != 'all' && !loading">
-        <v-row v-if="!loading">
-          <v-col cols="12" md="9">
-            <v-text-field
-                v-model.lazy="search"
-                append-icon="mdi-magnify"
-                label="Search"
-                outlined
-                dense
-                class="cyan lighten-5"
-                hide-details
-            ></v-text-field>
-          </v-col>
-          <v-col cols="12" md="3">
-            <v-menu
-                v-if="type == 'monthly'"
-                ref="menu"
-                v-model="menu"
-                :close-on-content-click="false"
-                :return-value.sync="month"
-                transition="scale-transition"
-                offset-y
-                min-width="290px"
-            >
-                <template v-slot:activator="{ on, attrs }">
-                <v-text-field
-                    v-model="month"
-                    placeholder="Select Month"
-                    prepend-inner-icon="mdi-calendar"
-                    readonly
-                    outlined
-                    hide-details
-                    dense
-                    color="info"
-                    v-bind="attrs"
-                    v-on="on"
-                ></v-text-field>
-                </template>
-                <v-date-picker
-                    v-model="month"
-                    type="month"
-                    color="info"
-                    no-title
-                    scrollable
-                >
-                <v-spacer></v-spacer>
-                <v-btn
-                    text
-                    color="primary"
-                    @click="menu = false"
-                >
-                    Cancel
-                </v-btn>
-                <v-btn
-                    text
-                    color="primary"
-                    @click="$refs.menu.save(month)"
-                >
-                    OK
-                </v-btn>
-                </v-date-picker>
-            </v-menu>
-            <v-autocomplete outlined v-model="year" hide-details placeholder="Select Year" v-if="type == 'yearly'" dense :items="[2020, 2021]"></v-autocomplete>
-          </v-col>
-        </v-row>
-        <v-row v-if="!loading && documents.length" class="my-3">
-          <v-col cols="12" sm="6" md="3">
-            <v-card flat color="info">
-              <v-card-text>
-                <v-row align="center" class="white--text">
-                  <v-col cols="8">
-                    <h1 class="mb-2">{{getAllDocuments}}</h1>
-                    <span>Uploaded</span>
-                  </v-col>
-                  <v-col cols="4">
-                    <v-icon size="50" color="info lighten-5">mdi-cloud-upload-outline</v-icon>
-                  </v-col>
-                </v-row>
-              </v-card-text>
-            </v-card>
-          </v-col>
-          <v-col cols="12" sm="6" md="3">
-            <v-card flat color="info">
-              <v-card-text>
-                <v-row align="center" class="white--text">
-                  <v-col cols="8">
-                    <h1 class="mb-2">{{getAllProcessedDocuments}}</h1>
-                    <span>Processed</span>
-                  </v-col>
-                  <v-col cols="4">
-                    <v-icon size="50" color="info lighten-5">mdi-check-circle</v-icon>
-                  </v-col>
-                </v-row>
-              </v-card-text>
-            </v-card>
-          </v-col>
-          <v-col cols="12" sm="6" md="3">
-            <v-card flat color="info">
-              <v-card-text>
-                <v-row align="center" class="white--text">
-                  <v-col cols="8">
-                    <h1 class="mb-2">{{getAllUnProcessedDocuments}}</h1>
-                    <span>Un Processed</span>
-                  </v-col>
-                  <v-col cols="4">
-                    <v-icon size="50" color="info lighten-5">mdi-close-circle</v-icon>
-                  </v-col>
-                </v-row>
-              </v-card-text>
-            </v-card>
-          </v-col>
-          <v-col cols="12" sm="6" md="3">
-            <v-card flat color="info">
-              <v-card-text>
-                <v-row align="center" class="white--text">
-                  <v-col cols="8">
-                    <h1 class="mb-2">{{getAllIndividualClients}}</h1>
-                    <span>Individual Clients</span>
-                  </v-col>
-                  <v-col cols="4">
-                    <v-icon
-                      size="50"
-                      color="info lighten-5"
-                    >
-                      mdi-account-multiple
-                    </v-icon>
-                  </v-col>
-                </v-row>
-              </v-card-text>
-            </v-card>
-          </v-col>
-        </v-row>
-        <v-data-table
-          v-if="!loading && documents.length"
-          :headers="headers"
-          :items="documents"
-          item-key="id"
-        >
-          <template v-slot:item.client_name="{ item }">
-            <span class="underline info--text link" @click="$router.push('/client/'+item.client_id+'/documents')"><strong>{{item.companyName}}</strong></span>
-          </template>
-          <template v-slot:item.total="{ item }">
-            {{getTotalDocs(item.client_id)}}
-          </template>
-          <template v-slot:item.processed="{ item }">
-            {{getProcessedDocs(item.client_id)}}
-          </template>
-          <template v-slot:item.unprocessed="{ item }">
-            {{getUnProcessedDocs(item.client_id)}}
-          </template>
-        </v-data-table>
-      </div>
-
-      <div v-if="type == 'all' && !loading">
-        <AllDocuments />
-      </div>
-
       <v-row v-if="!loading && !documents.length" class="my-5">
         <v-col cols="12" class="text-center">
             <p class="mt-3 info--text">No Document Found</p>
         </v-col>
       </v-row>
+
+      <v-row v-if="!loading && documents.length">
+        <v-col cols="12">
+          <v-text-field
+              v-model="search"
+              append-icon="mdi-magnify"
+              label="Search"
+              outlined
+              dense
+              class="cyan lighten-5"
+              single-line
+              hide-details
+          ></v-text-field>
+        </v-col>
+
+        <v-col cols="12" class="d-flex" v-if="selected.length">
+          <v-btn color="info" text :disabled="isStatusUpdating" @click="archiveDocuments()"><v-icon>mdi-arrow-down</v-icon> Archive</v-btn>
+          <download-excel
+            class="btn btn-default"
+            :data="selected"
+            :fields="json_fields"
+            worksheet="Archived Documents Worksheet"
+            name="Archived Documents.xls"
+          >
+            <v-btn color="info" text :disabled="isStatusUpdating">Export</v-btn>
+          </download-excel>
+          <v-btn color="info" text :disabled="isStatusUpdating" @click="updateStaus('Processed')">Mark as processed</v-btn>
+          <v-btn color="info" text :disabled="isStatusUpdating" @click="updateStaus('Unprocessed')">Mark as unprocessed</v-btn>
+        </v-col>
+      </v-row>
+
+      <v-data-table
+        v-if="!loading && documents.length"
+        v-model="selected"
+        :headers="headers"
+        :items="documents"
+        :search="search"
+        :single-select="singleSelect"
+        item-key="id"
+        show-select
+      >
+        <template v-slot:item.file_name="{ item }">
+            <a :href="item.file_path" target="_blank"><strong>{{item.file_name}}</strong></a>
+        </template>
+      </v-data-table>
     </v-container>
+
+    <v-dialog v-model="addModal" title="Upload Document" max-width="500" :scrollable="false" persistent>
+      <v-card>
+          <v-card-title class="d-flex">
+              <h3 class="info--text">Upload Document</h3>
+              <v-icon @click="addModal = !addModal" class="ml-auto">mdi-close</v-icon>
+          </v-card-title>
+          <v-card-text class="mt-4">
+            <v-divider></v-divider>
+              <v-row class="mt-4">
+                  <v-col cols="12">
+                      <v-autocomplete :items="getUsers" item-text="name" return-object v-model="document.client" placeholder="Select Client" color="info" outlined hide-details></v-autocomplete>
+                  </v-col>
+                  <v-col cols="12">
+                      <v-text-field  v-model="document.category" hide-details placeholder="Category" color="info" outlined></v-text-field>
+                  </v-col>
+                  <v-col cols="12">
+                      <v-select :items="['Processed', 'Unprocessed']"  v-model="document.status" hide-details placeholder="Status" color="info" outlined></v-select>
+                  </v-col>
+                  <v-col cols="12">
+                      <v-menu
+                          ref="menu"
+                          v-model="menu"
+                          :close-on-content-click="false"
+                          :return-value.sync="document.month"
+                          transition="scale-transition"
+                          offset-y
+                          min-width="290px"
+                      >
+                          <template v-slot:activator="{ on, attrs }">
+                          <v-text-field
+                              v-model="document.month"
+                              placeholder="Select Month"
+                              prepend-inner-icon="mdi-calendar"
+                              readonly
+                              outlined
+                              hide-details
+                              color="info"
+                              v-bind="attrs"
+                              v-on="on"
+                          ></v-text-field>
+                          </template>
+                          <v-date-picker
+                              v-model="document.month"
+                              type="month"
+                              color="info"
+                              no-title
+                              scrollable
+                          >
+                          <v-spacer></v-spacer>
+                          <v-btn
+                              text
+                              color="primary"
+                              @click="menu = false"
+                          >
+                              Cancel
+                          </v-btn>
+                          <v-btn
+                              text
+                              color="primary"
+                              @click="$refs.menu.save(document.month)"
+                          >
+                              OK
+                          </v-btn>
+                          </v-date-picker>
+                      </v-menu>
+                  </v-col>
+                  <v-col cols="12">
+                      <v-file-input
+                        outlined
+                        color="info"
+                        prepend-inner-icon="mdi-file"
+                        @change="getPickedDocument"
+                        label="Upload File"
+                      ></v-file-input>
+                  </v-col>
+                  <v-col cols="12">
+                      <v-btn block depressed color="info" @click="submitDocument" :disabled="createLoading">{{createLoading ? 'Uploading Document...' : 'Upload Document'}}</v-btn>
+                  </v-col>
+              </v-row>
+          </v-card-text>
+      </v-card>
+  </v-dialog>
   </div>
 </template>
 
 <script>
 // import moment from 'moment'
 import {mapActions, mapGetters} from 'vuex'
-import AllDocuments from '@/components/Documents/AllDocuments'
 export default {
   name: "Documents",
-  components: {
-    AllDocuments
-  },
   computed: {
     ...mapGetters(["getUsers", "getDocuments", "getToast"]),
-    getAllDocuments() {
-      return this.clientDocs.length
-    },
-    getAllProcessedDocuments() {
-      const docs = this.clientDocs.filter(doc => doc.status == 'Processed')
-      return docs.length
-    },
-    getAllUnProcessedDocuments() {
-      const docs = this.clientDocs.filter(doc => doc.status == 'Unprocessed')
-      return docs.length
-    },
-    getAllIndividualClients() {
-      return this.documents.length
-    }
   },
   methods: {
     ...mapActions(['createDocument', 'fetchUsers', 'fetchDocuments', 'updateDocumentStatus', 'updateArchive']),
-    async setClientDocuments() {
-      const currentMonth = this.month.split('-')[1]
-      const currentYear = this.year
-      this.clientDocs = []
-      if(this.type == 'all') {
-        this.documents = this.getDocuments
-        this.clientDocs = this.getDocuments
-
+    getPickedDocument(e) {
+      this.document.file.file_name = e.name
+      this.document.file.file = e
+    },
+    submitDocument() {
+      let documentData = Object.assign({}, this.document)
+      if(!documentData.client) {
+        this.$store.commit('setToast', {message: "Please Select Client", color: 'red', show: true})
         return
       }
-      await this.getDocuments.forEach(item => {
-        let itemMonth = item.month.split('-')[1]
-        let itemYear = item.month.split('-')[0]
-        if(this.type == 'monthly') {
-          if((itemMonth == currentMonth) && !item.isArchived) {
-            this.clientDocs.push(item)
-          }
-        }
-        if(this.type == 'yearly') {
-          if((itemYear == currentYear) && !item.isArchived) {
-            this.clientDocs.push(item)
-          }
-        }
-        
-      })
-      this.documents = []
-      if(this.clientDocs.length) {
-        await this.clientDocs.forEach(item => {
-          let doc = this.documents.find(cd => cd.client_id == item.client_id)
-          if(!doc) {
-            this.documents.push(item)
-          }
-        })
+      if(!documentData.category || documentData.category == "") {
+        this.$store.commit('setToast', {message: "Please Enter Category", color: 'red', show: true})
+        return
       }
+      if(!documentData.month || documentData.month == "") {
+        this.$store.commit('setToast', {message: "Please Select Month", color: 'red', show: true})
+        return
+      }
+      if(!documentData.status) {
+        this.$store.commit('setToast', {message: "Please Select Status", color: 'red', show: true})
+        return
+      }
+      if(!documentData.file.file) {
+        this.$store.commit('setToast', {message: "Please Pick Document File", color: 'red', show: true})
+        return
+      }
+      documentData.client_name = documentData.client.name
+      documentData.client_id = documentData.client.id
+      delete documentData.client
+      this.createLoading = true
+      this.createDocument(documentData)
+    },
+    async updateStaus(status) {
+      Promise.all(
+        this.selected.map(async doc => {
+            this.isStatusUpdating = true
+            doc.status = status
+            await this.updateDocumentStatus(doc)
+        })
+      )
+    },
 
-      this.allDocs = this.documents
-    },
-    getTotalDocs(id) {
-      const total = this.clientDocs.filter(doc => doc.client_id == id)
-      return total.length
-    },
-    getProcessedDocs(id) {
-      const processed = this.clientDocs.filter(doc => doc.client_id == id && doc.status == 'Processed')
-      return processed.length
-    },
-    getUnProcessedDocs(id) {
-      const unprocessed = this.clientDocs.filter(doc => doc.client_id == id && doc.status == 'Unprocessed')
-      return unprocessed.length
+    async archiveDocuments() {
+      console.log(this.selected)
+      Promise.all(
+        this.selected.map(async doc => {
+            this.isStatusUpdating = true
+            doc.isArchived = true
+            await this.updateArchive(doc)
+        })
+      )
     },
   },
   created() {
@@ -286,69 +231,71 @@ export default {
       this.fetchDocuments()
     }
     else {
-      this.setClientDocuments()
+      this.documents = this.getDocuments.filter(doc => !doc.isArchived)
+      // Sorting in Ascending Order
+      this.documents = this.documents.sort((a,b) => new Date(a.month) - new Date(b.month))
+      this.json_data = this.documents
     }
   },
   watch: {
     getDocuments(val) {
       if(val) {
+        this.createLoading = false
         this.loading = false
-        this.setClientDocuments()
+        this.documents = val.filter(doc => !doc.isArchived)
+        // Sorting in Ascending Order
+        this.documents = this.documents.sort((a,b) => new Date(a.month) - new Date(b.month))
+        this.json_data = this.documents
+        this.addModal = false
       }
     },
     getToast(val) {
       if(val) {
+        this.createLoading = false
         this.loading = false
+        this.isStatusUpdating = false
       }
-    },
-    search(val) {
-      if(val) {
-        console.log(val)
-        this.documents = this.allDocs.filter(doc => doc.companyName.toLowerCase().indexOf(val.toLowerCase()) >= 0)
-      }
-      else {
-        this.documents = this.allDocs
-      }
-    },
-    month() {
-      this.setClientDocuments()
-    },
-    year() {
-      this.setClientDocuments()
     }
   },
   data() {
     return {
       menu: false,
-      type: 'all',
-      month: new Date().getFullYear()+"-"+new Date().getMonth()+1,
-      year: new Date().getFullYear(),
       addModal: false,
       search: null,
+      singleSelect: false,
+      selected: [],
       loading: false,
+      createLoading: false,
+      isStatusUpdating: false,
+      document: {
+        client: null,
+        category: null,
+        month: null,
+        status: null,
+        isArchived: false,
+        file: {
+          file_name: null,
+          file: null,
+        },
+      },
       headers: [
-        { text: 'Client', sortable: true, value: 'client_name'},
-        { text: 'Total', value: 'total', sortable: true },
-        { text: 'Processed', value: 'processed', sortable: true },
-        { text: 'Unprocessed', value: 'unprocessed', sortable: true }
+        { text: 'File Name', sortable: true, value: 'file_name'},
+        { text: 'Client', value: 'client_name', sortable: true },
+        { text: 'Category', value: 'category', sortable: true },
+        { text: 'Month', value: 'month', sortable: true },
+        { text: 'Status', value: 'status', sortable: true },
       ],
+      json_data: [],
       documents: [],
-      clientDocs: [],
-      allDocs: []
+      json_fields: {
+        "File Name": "file_name",
+        "File Path": "file_path",
+        "Client Name": "client_name",
+        "Category": "category",
+        "Month": "month",
+        "Status": "status"
+      }
     }
   }
 };
 </script>
-
-<style scoped>
-.active {
-  padding-bottom: 2px;
-  border-bottom: 2px solid skyblue;
-}
-.link:hover {
-  cursor: pointer;
-}
-.underline {
-  text-decoration: underline;
-}
-</style>
