+++ src/views/Login.vue
@@ -30,15 +30,13 @@
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
+import firebase from 'firebase';
+require('firebase/auth');
+import router from '../router/index';
+
 export default {
-    name: "Login",
-    computed: {
-        ...mapGetters(["getToast"])
-    },
     data(){
         return{
-            loading: false,
             email: '',
             password: '',
             error: '',
@@ -50,25 +48,15 @@ export default {
             ]
         }
     },
-    watch: {
-        getToast(val) {
-            if(val) {
-                this.loading = false
-            }
-        }
-    },
-    created() {
-        if(this.getLoggedUser()) {
-            this.$router.push('/')
-        }
-    },
     methods: {
         async pressed() {
-            const payload = {
-                email: this.email,
-                password: this.password
+            try{
+                const val = await firebase.auth().signInWithEmailAndPassword(this.email, this.password);
+                console.log('loggedin user: ' + val);
+                router.push('/');
+            }catch(err){
+                console.log(err);
             }
-            this.$store.dispatch('loginUser', payload)
         }
     }
 }
